Used libcuckoo concurrent hash map Used concurrentqueue from https://github.com/cameron314/concurrentqueue Added support for both autoconf and cmake Fixed multiple bugs/issues

+++ web/default/src/DefaultOAUTHController.cpp
@@ -61,10 +61,10 @@ bool DefaultOAUTHController::service(HttpRequest* req, HttpResponse* res)
 			temps.push_back((it->first + "=" + it->second));
 		}
 	}
-	for (int var = 0; var < temps.size(); ++var)
+	for (int var = 0; var < (int)temps.size(); ++var)
 	{
 		allparst += temps.at(var);
-		if(var!=temps.size()-1)
+		if(var!=(int)temps.size()-1)
 			allparst += "&";
 	}
 	std::cout << "req and oauth params" <<std::endl;
